fix(ci): use short sha for mr builds
This is a workaround. ArgoCD smart truncates the {{ branch_slug }} to 50 characters using gosimple/slug, GitLab truncates the \ to 63, leading to images with tags longer than 50 characters. But since ArgoCD truncates to 50, we sometimes get an ImagePullBackOff.

To circumvent this, we could use the newish go templating support in ArgoCD, but the function (slugify) isn't stable yet. Revert this commit once it is and our operators is updated.