feat(KSTIP-1029): Removed unnecessary tests and added ones to check Gesuchstatus
The tests for finding Gesuche with no formular and with no person in Ausbildung have been removed as they were found to be unnecessary. Additionally, the option to update person created Zuordnung was commented out for the time being. New tests have been added to verify that the Gesuchstatus is not IN_BEARBEITUNG_GS or KOMPLETT_EINGEREICHT for any Gesuch returned by certain GesuchApiSpec methods.