feat(KSTIP-910): Add new user welcome email functionality
Added the ability to send a welcome email to a newly created user. This feature includes the necessary classes, methods, and API specifications. It also includes testing functionality for the new feature. This will facilitate the onboarding process for new users by giving them a warm welcome and instructions for next steps.